pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "LitCoach API"
authors = [
    { name = "Reza Tabrizi", email = "rtabrizi03@gmail.com" }
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.8.0",
    "certifi==2024.12.14",
    "charset-normalizer==3.4.1",
    "click==8.1.8",
    "dnspython==1.16.0",
    "fastapi==0.115.6",
    "h11==0.14.0",
    "idna==3.10",
    "pydantic==2.10.5",
    "pydantic-settings==2.7.1",
    "pydantic-core==2.27.2",
    "pymongo==3.12.0",
    "python-dotenv==1.0.1",
    "requests==2.32.3",
    "sniffio==1.3.1",
    "starlette==0.41.3",
    "typing-extensions==4.12.2",
    "urllib3==2.3.0",
    "uvicorn==0.34.0",
    "tiktoken (>=0.8.0,<0.9.0)",
    "openai (>=1.60.2,<2.0.0)"
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "api" }
]
